IA/requirements.txt
flask
tensorflow==2.11.0
opencv-python
numpy<2.0
tqdm
scikit-learn
matplotlib
Pillow